provide instructions to retrieve the password
Allow kubernetes/kubectl to reveal the password instead of embedding
it in NOTEs.txt (which is available in `helm get notes <release name>`).

This works for both values provided in values.yaml and anything
that is auto-generated. Helm templates all of the files at once,
before applying the resources, so we cannot lookup in NOTES what
was generated in another file.